requirements.txt
absl-py==0.14.1
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
cachetools==4.2.4
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
entrypoints==0.3
fire==0.4.0
fsspec==2021.10.0
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
idna==3.2
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.5
jupyter-core==4.8.1
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
matplotlib-inline==0.1.3
multidict==5.1.0
nest-asyncio==1.5.1
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
protobuf==3.18.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytorch-lightning==1.4.9
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
torch==1.9.1
torchmetrics==0.5.1
torchvision==0.10.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.1
yarl==1.6.3
